proxy: generate signed signout url for authenticate dashboard
When user click to "Sign Out" button in dashboard page:

 - A request is sent to proxy
 - Proxy generates a signed signout url, then redirect to authenticate
 service to sign user out.

But when user access the dashboard page in authenticate service itself,
the signed signout url is not generated, because the request is not go
through proxy, but handled by the authenticate service itself.

To fix it, we will generate the signed signout url if user access
dashboard directly in authenticate service.

Fixes #867