cache: fix data race in NotifyJoin
In 35af5c0b91e1f93313a6ca67ed62cd1a72b11c5b, the check for multiple
cache servers in NotifyJoin is made to be done in a goroutine. That can
lead to a data race, because the memberlist can be changed at the time
the goroutine was run. go warns about this race when test memberlist was
run with "-race".

To fix this, we pass the nil check as argument to goroutine.