pkg/storage: introduce storage.Backend Sync method
Currently, we're doing "sync" in databroker server. If we're going to
support multiple databroker servers instance, this mechanism won't work.

This commit moves the "sync" to storage backend, by adding new Sync
method. The Sync method will return a chanel for the caller. Everytime
something happens inside the storage, we notify the caller by sending a
message to the channel.