config: separate wildcard hosts where possible
Envoy supports virtual hosts with wildcard domains, as long as the
wildcard token appears at the very beginning or the end of the domain
(that is, Envoy supports suffix and prefix domain wildcards). Update the
configuration builder logic to take advantage of this support whenever
possible, using the current regex route matching as a fallback.

This may improve routing performance in some cases, by reducing the size
of the routing table for the catch-all virtual host.

Add a utility method isSpecialWildcardHost() to determine whether a host
has a wildcard token that is not at the very beginning or end. Update
the existing route configuration test to add one of these special
wildcard hosts.