integration tests: bring templates in sync (#4230)
There are a few references in the integration test configuration
templates to files that are not currently present in the v0.17 branch.
There is also a discrepancy between the templates and configuration
output with regard to the default Pomerium image tag name (master/main).

Add the missing files (from the 'main' branch) and update the default
image tag. (The corresponding changes in the 'main' branch were included
in commit 9980fd6141b776a1a8599f5f0ee2428d223f0015.)

With these changes, running

    go run ./integration/cmd/pomerium-integration-tests/ generate-configuration

should produce no changes in the integration/clusters/ configuration
output files.