fix: Remove default prompt
There's a difference in Azure if you remove the prompt parameter altogether (transparently redirects user in case user already has a valid session, but presents the user with a login screen in case the user doesn't). This is by default a good approach, also being used by default Azure apps, such as Azure AD or the Azure Portal. This change opens up the possibility for transparent login with Azure, but still allows for customisation of the prompt parameter.