authorize: compute "real" IP from X-Forwarded-For
WIP -- this doesn't work correctly yet

Update the authorize log IP address field to match the access log IP
address field.

Add a new method getIPAddress() to compute the "real" client IP address
based on the contents of the X-Forwarded-For header and the
XffNumTrustedHops and SkipXffAppend configuration option.

This is intended to match downstream_remote_address from the access log:
https://www.envoyproxy.io/docs/envoy/latest/api-v3/data/accesslog/v3/accesslog.proto#envoy-v3-api-field-data-accesslog-v3-accesslogcommon-downstream-remote-address