forward proxy proof of concept
Envoy configuration: create a duplicate copy of the 'http-ingress'
listener but configured as an "internal listener." Add a corresponding
cluster and filter chain to accept CONNECT requests and route them to
this internal listener.

I **think** this may allow Pomerium to function as a forward proxy,
suitable for use via PAC (proxy auto-configuration).

For now, the new filter chain has the hard-coded domain
"forward-proxy.localhost.pomerium.io", but this would probably be
controlled by a new configuration setting.