Sanity-check file timestamps before polling
Before adding files to the polling watcher, ensure that the underlying
filesystem provides functional timestamps. Since changes are detected by
comparing file modification times, incorrect values can cause the polling
file watcher to behave erratically.

This is done by checking to see if we can repeatedly stat() the file
and get the same modification timestamp twice in a row. If repeated
calls to stat() each return different timestamps (for instance, if the
filesystem returns 'now' as the modification time on each request),
the file will not be polled and a warning will be logged.

This check will attempt to stat() each file up to 10 times. On
functioning filesystems, this will result in only one extra
stat() per file.