envoy: dynamic access log fields
This adds the ability to include any available field in the envoy access
logs, in addition to the built-in field names we have predefined.

Entries in the access_log_fields list in the config that have the form
'name=path', where path is a protopath string with an implied root of
type envoy.data.accesslog.v3.HTTPAccessLogEntry, will add fields to all
access log events. These fields will use the given name as the key, and
use the path to look up the value in the log entry sent by envoy.

Paths can be used to look up fields deeply nested in the message,
as well as access dynamic metadata (with google.protobuf.Any values),
specific entries in arrays or maps, or select entire message fields
to log as a json object.