Makefile: unset GOOS for get-envoy target
The build-dev-docker.sh script will run

  env GOOS=linux make build

to build Pomerium for Linux (before copying this binary into a minimal
Docker image).

However, now that the get-envoy target is implemented using a Go
command, it too will be affected by GOOS. As a result running this step
on macOS will result in an error when trying to execute a Linux
get-envoy binary.

Instead, let's unset GOOS when running the get-envoy command, so that it
defaults to building for the host OS.