config: fix JWT groups filter option
When applying the settings proto, update the JWT groups filter option
only if the filter set is non-empty.

This is important when Pomerium via the Ingress Controller in
combination with Pomerium Enterprise. In this scenario there is a
settings proto applied from both Ingress Controller and the Enterprise
console, and we want to make sure the one from Ingress Controller does
not overwrite the filter settings from Enterprise.