config: migrate remaining mapstructure imports
The github.com/mitchellh/mapstructure repo is archived, and
github.com/go-viper/mapstructure is the blessed replacement according to
https://gist.github.com/mitchellh/90029601268e59a29e64e55bab1c5bdc#the-list.

Update the remaining mapstructure imports to the new package.

The behavior of StringToSliceHookFunc() has changed slightly, in a way
that breaks decoding of the POLICY environment variable. However we can
use StringToWeakSliceHookFunc() instead, which provides the old
behavior. Other than that, it looks like all the unit tests still pass.