policy: share input type with criteria tests
We have a duplicate set of types for the Rego input to policy
evaluation: one in the authorize/evaluator package and one in the
pkg/policy/criteria tests. (This was done to avoid an import cycle
between the two packages.)

However, it presents the possibility for these two to get out of sync,
which would reduce the fidelity of the criteria unit tests. (In fact,
this has already happened to a very minor extent: in one version the
Headers field is a map[string]string while in the other it is a
map[string][]string.)

Let's consolidate these to one set of types, in a new package.