hold back go-proxyproto, update protovalidate-go
The go-proxyproto module appears to have an incorrect go directive, so
hold off on this version update for now.

The bufbuild/protovalidate/protocolbuffers/go module requires a newer
version of the bufbuild/protovalidate-go module. This also introduces a
small formatting change to the validation error message in one of our
unit tests.